# grpc-portal/cmd/proto/user.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "cmd/proto/user.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "UsersService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/users": {
      "post": {
        "summary": "Create a new user",
        "operationId": "CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoUserResponse"
            }
          },
          "201": {
            "description": "User successfully created",
            "schema": {
              "$ref": "#/definitions/protoUserResponse"
            }
          },
          "400": {
            "description": "Bad request - Invalid input",
            "schema": {}
          },
          "500": {
            "description": "Internal server error",
            "schema": {}
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoUserRequest"
            }
          }
        ],
        "tags": [
          "Users"
        ]
      }
    }
  },
  "definitions": {
    "protoUserRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "User's name"
        },
        "country": {
          "type": "string",
          "description": "Country ISO"
        },
        "email": {
          "type": "string",
          "description": "User's email"
        },
        "password": {
          "type": "string",
          "description": "User's password"
        },
        "contact": {
          "type": "string",
          "description": "User's contact number"
        },
        "address": {
          "type": "string",
          "description": "User's address"
        },
        "state": {
          "type": "string",
          "description": "User's state"
        },
        "role": {
          "type": "string",
          "description": "User's role"
        }
      }
    },
    "protoUserResponse": {
      "type": "object",
      "properties": {
        "uuid": {
          "type": "string",
          "description": "User UUID"
        },
        "name": {
          "type": "string",
          "description": "User's name"
        },
        "email": {
          "type": "string",
          "description": "User's email"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}